mega_test1_s3_bucket/requirements.txt
Flask==2.2.5
matplotlib==3.3.4
speedtest-cli==2.1.3

